requirements.txt
watchdog>=3.0.0
openai-whisper>=20231117
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.5.0
